Script for setting up language subtag collection
This script will parse the IANA language subtag registry file
(which should be downloaded to language-subtag-registry.txt in the
same directory as this script) and populate a MongoDB collection with
the "language" and "extlang" records from the registry. The script is
intended to be run at the server's command line, like this:

    php setup-language-registry.php dbname

where dbname is the name of the Mongo database (e.g., "languageforge").