WIP: select language only at project creation
Since the language (or rather, its 2- or 3-letter code) is part of
the project code from which the database (and repository) name is
derived, we should not allow a project's language to be changed
after creation. Instead, we should ask the user to choose a language
during project creation only, and not change it thereafter.