Fix bug with typeahead invocation
Since the typeahead directive calls its provided search() function
with a parameter list {term: some_value}, the parameter provided in
the HTML invocation of the typeahead directive apparently needs to
be "term", and not any other parameter name. Before this fix, the
search term was "lagging" one character behind: e.g., for a project
in English, typing "engli" into the Project Language input field
would trigger a search for "engl" in the languages list. Typing the
next letter, "s", would search for "engli" instead of "englis". And
so on. With this fix, the search term is now exactly what was typed.