Fix too-long lines around user avatar "button"
The user avatar "button" in the header had lines that were slightly
too long to its left and right. I finally tracked this error down to
its source. Using "position: relative" was correcting the positioning
of the element, but leaving a space for it in the "original" location,
which was in turn causing its containing div to be a bit too large.