Submenus for buttons that need them
Still need to tweak submenu CSS a bit to make it look better, but
this is a good start. Also need to implement some Angular animation
to make it all look good: the chevron should flip, the submenu should
fade (or slide) into view and away, and the menu should go away once
it hasn't been hovered on for a short time.