Added sample data from Thai food dictionary
For the "Add missing information" and "Gather words" submenus, it will
make the most sense to work with some real(-ish) data, so I've added a
few entries from the Thai food dictionary as a good starting point. This
sample data isn't yet used, but pretty soon we'll have the lexService
functions returning the sample data, or specific subsets of it, so that
the forms we're designing can have some real data in them.

Once the database backend is written, of course, these services will be
updated to work with the real database.