Improve submenus: correct mouse pointer, underline
Submenu "links" (that is, the word clicked on to show the submenu)
should use the "hand" mouse pointer, but not show an underline when
hovered over because they don't go to a new page. The links that do go
to a new page, though, should still show an underline when hovered over.

With this commit, the submenus look the way I want them to look.